example: compiler warning fixes

+++ example/direct_tcpip.c
@@ -286,7 +286,8 @@ int main(int argc, char *argv[])
                     continue;
                 }
                 if(nwritten < 0) {
-                    fprintf(stderr, "libssh2_channel_write: %zd\n", nwritten);
+                    fprintf(stderr, "libssh2_channel_write: %d\n",
+                            (int)nwritten);
                     goto shutdown;
                 }
                 wr += nwritten;
@@ -297,7 +298,7 @@ int main(int argc, char *argv[])
             if(LIBSSH2_ERROR_EAGAIN == len)
                 break;
             else if(len < 0) {
-                fprintf(stderr, "libssh2_channel_read: %zd", len);
+                fprintf(stderr, "libssh2_channel_read: %d", (int)len);
                 goto shutdown;
             }
             wr = 0;
