+++ example/tcpip-forward.c
@@ -273,7 +273,8 @@ int main(int argc, char *argv[])
             do {
                 nwritten = libssh2_channel_write(channel, buf, len);
                 if(nwritten < 0) {
-                    fprintf(stderr, "libssh2_channel_write: %zd\n", nwritten);
+                    fprintf(stderr, "libssh2_channel_write: %d\n",
+                            (int)nwritten);
                     goto shutdown;
                 }
                 wr += nwritten;
@@ -285,7 +286,8 @@ int main(int argc, char *argv[])
             if(LIBSSH2_ERROR_EAGAIN == len)
                 break;
             else if(len < 0) {
-                fprintf(stderr, "libssh2_channel_read: %zd", len);
+                fprintf(stderr, "libssh2_channel_read: %d",
+                        (int)len);
                 goto shutdown;
             }
             wr = 0;
